jobque: 64-bit Stream byte-buffer API (push / pushBatch / pop)
The byte-buffer Stream API truncated 64-bit sizes/counts at the boundary even
though the pipe internals (deque<vector<uint8_t>>) are already size_t: Stream::push
took uint32_t size, Stream::pushBatch took const uint32_t* sizes + int count,
streamPushBatch built a vector<uint32_t> and passed (int)data.size, and the
pop-side invoke_stream_block took uint32_t size with (uint32_t)item.size() casts.

Widen the Stream API to 64-bit: push size -> uint64_t; pushBatch sizes ->
const uint64_t*, count -> int64_t; invoke_stream_block size -> uint64_t (drop the
(uint32_t) casts); streamPushBatch -> vector<uint64_t> sizes + int64_t(data.size);
aot_builtin_jobque.h declarations updated to match. The das-bound functions
(streamPush / streamPushBatch / streamPop) are unchanged -- they take TArray and
call the internal Stream methods -- so no das-API or AOT-prototype change.

A >4GB-buffer / >2G-batch repro isn't CI-feasible, so no new test (same as the
foreach/fio PRs); validated tests/jobque/test_jobque_stream.das (20/20) + full suite.

Followup to #3175 (64-bit container sweep). Left as separate follow-ups: the
int-bounded Stream::total() queue-depth accessor and the Channel typed-message
push/pushBatch path.

+++ include/daScript/simulate/aot_builtin_jobque.h
@@ -142,8 +142,8 @@ namespace das {
         Stream() { mTrackMagic = TRACK_STREAM; }
         Stream( int count ) { mTrackMagic = TRACK_STREAM; mRemaining = count; }
         virtual ~Stream();
-        void push ( const uint8_t * data, uint32_t size );
-        void pushBatch ( const uint8_t * const * data, const uint32_t * sizes, int count );
+        void push ( const uint8_t * data, uint64_t size );
+        void pushBatch ( const uint8_t * const * data, const uint64_t * sizes, int64_t count );
         void pop ( const TBlock<void, TTemporary<TArray<uint8_t> const>> & blk, Context * context, LineInfoArg * at );
         bool tryPop ( const TBlock<void, TTemporary<TArray<uint8_t> const>> & blk, Context * context, LineInfoArg * at );
         bool popWithTimeout ( int timeoutMs, const TBlock<void, TTemporary<TArray<uint8_t> const>> & blk, Context * context, LineInfoArg * at );
@@ -155,7 +155,7 @@ namespace das {
             lock_guard<mutex> guard(mCompleteMutex);
             for ( auto & v : pipe ) {
                 Array arr;
-                array_mark_locked(arr, (void *)v.data(), (uint32_t)v.size());
+                array_mark_locked(arr, (void *)v.data(), v.size());
                 tt(&arr);
             }
         }
@@ -164,7 +164,7 @@ namespace das {
             lock_guard<mutex> guard(mCompleteMutex);
             for ( auto & v : pipe ) {
                 Array arr;
-                array_mark_locked(arr, (void *)v.data(), (uint32_t)v.size());
+                array_mark_locked(arr, (void *)v.data(), v.size());
                 tt(&arr);
             }
             pipe.clear();
